test get/select docs with list values
